+++ PythonAPI/carla/source/carla/__init__.py
@@ -6,3 +6,8 @@
 
 # pylint: disable=W0401,import-self
 from .libcarla import *
+
+# Allow from carla.command import ...
+import sys
+sys.modules["carla.command"] = command
+del sys